+++ tests/conftest.py
@@ -4,6 +4,7 @@ import pytest
 
 
 pytest.register_assert_rewrite("optuna.testing.pytest_samplers")
+pytest.register_assert_rewrite("optuna.testing.pytest_importance")
 pytest.register_assert_rewrite("optuna.testing.pytest_storages")
